+++ src/cpx_io/cpx_system/cpx_e.py
@@ -123,7 +123,7 @@ class CpxE(CpxBase):
 
         data &= ~self._control_bit_value
         value = self.read_reg_data(*_ModbusCommands.data_system_table_read)
-        return value
+        return value[0]
 
     def module_count(self) -> int:
         """returns the total count of attached modules"""
@@ -146,7 +146,7 @@ class CpxE(CpxBase):
     def read_device_identification(self) -> int:
         """returns Objects IDO 1,2,3,4,5"""
         data = self.read_function_number(43)
-        return data[0]
+        return data
 
     def read_module_count(self) -> int:
         """Reads and returns IO module count as integer"""
@@ -289,7 +289,7 @@ class CpxE8Do(_CpxEModule):
         the error will be sent to the bus module and displayed on the module by the error LED.
         """
         function_number = 4828 + 64 * self.position + 0
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register
         if short_circuit is None:
@@ -315,7 +315,7 @@ class CpxE8Do(_CpxEModule):
         the corresponding output must be reset and to restore the power.
         """
         function_number = 4828 + 64 * self.position + 1
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register
         if value:
@@ -367,7 +367,7 @@ class CpxE16Di(_CpxEModule):
         the error will be sent to the bus module and displayed on the module by the error LED.
         """
         function_number = 4828 + 64 * self.position + 0
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register
         if value:
@@ -388,7 +388,7 @@ class CpxE16Di(_CpxEModule):
         the CPX-E automation system must be switched off and on to restore the power.
         """
         function_number = 4828 + 64 * self.position + 1
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register
         if value:
@@ -411,7 +411,7 @@ class CpxE16Di(_CpxEModule):
             raise ValueError("Value {value} must be between 0 and 3")
 
         function_number = 4828 + 64 * self.position + 1
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register, delete bit 4+5 from it
         # and refill it with value
@@ -432,7 +432,7 @@ class CpxE16Di(_CpxEModule):
             raise ValueError("Value {value} must be between 0 and 3")
 
         function_number = 4828 + 64 * self.position + 1
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register, delete bit 6+7 from it
         # and refill it with value
@@ -497,8 +497,8 @@ class CpxE4AiUI(_CpxEModule):
 
         function_number = 4828 + 64 * self.position
 
-        reg_01 = self.base.read_function_number(function_number + 13)[0]
-        reg_23 = self.base.read_function_number(function_number + 14)[0]
+        reg_01 = self.base.read_function_number(function_number + 13)
+        reg_23 = self.base.read_function_number(function_number + 14)
 
         if channel == 0:
             function_number += 13
@@ -527,8 +527,8 @@ class CpxE4AiUI(_CpxEModule):
 
         function_number = 4828 + 64 * self.position
 
-        reg_01 = self.base.read_function_number(function_number + 15)[0]
-        reg_23 = self.base.read_function_number(function_number + 16)[0]
+        reg_01 = self.base.read_function_number(function_number + 15)
+        reg_23 = self.base.read_function_number(function_number + 16)
 
         if channel == 0:
             function_number += 15
@@ -564,7 +564,7 @@ class CpxE4AiUI(_CpxEModule):
         the error will be sent to the bus module and displayed on the module by the error LED.
         """
         function_number = 4828 + 64 * self.position + 0
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register
         if short_circuit == None:
@@ -587,7 +587,7 @@ class CpxE4AiUI(_CpxEModule):
         the CPX-E automation system must be switched off and on to restore the power.
         """
         function_number = 4828 + 64 * self.position + 1
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register
         if value:
@@ -605,7 +605,7 @@ class CpxE4AiUI(_CpxEModule):
         * True: Linear scaled
         """
         function_number = 4828 + 64 * self.position + 6
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register
         if value:
@@ -623,7 +623,7 @@ class CpxE4AiUI(_CpxEModule):
         The sensor supply can also be switched off and switched on during operation.
         """
         function_number = 4828 + 64 * self.position + 6
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register
         if value:
@@ -644,7 +644,7 @@ class CpxE4AiUI(_CpxEModule):
         displayed with the error LED on the module.
         """
         function_number = 4828 + 64 * self.position + 6
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register
         if value:
@@ -664,7 +664,7 @@ class CpxE4AiUI(_CpxEModule):
         the automation system CPX-E must be switched off and on to restore the power.
         """
         function_number = 4828 + 64 * self.position + 6
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register
         if value:
@@ -778,8 +778,8 @@ class CpxE4AoUI(_CpxEModule):
 
         function_number = 4828 + 64 * self.position
 
-        reg_01 = self.base.read_function_number(function_number + 11)[0]
-        reg_23 = self.base.read_function_number(function_number + 12)[0]
+        reg_01 = self.base.read_function_number(function_number + 11)
+        reg_23 = self.base.read_function_number(function_number + 12)
 
         if channel == 0:
             function_number += 11
@@ -810,7 +810,7 @@ class CpxE4AoUI(_CpxEModule):
         the error will be sent to the bus module and displayed on the module by the error LED.
         """
         function_number = 4828 + 64 * self.position + 0
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register
         if short_circuit is None:
@@ -839,7 +839,7 @@ class CpxE4AoUI(_CpxEModule):
         the automation system CPX-E must be switched off and on to restore the power.
         """
         function_number = 4828 + 64 * self.position + 1
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register
         if value:
@@ -860,7 +860,7 @@ class CpxE4AoUI(_CpxEModule):
         the automation system CPX-E must be switched off and on to restore the power.
         """
         function_number = 4828 + 64 * self.position + 1
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register
         if value:
@@ -877,7 +877,7 @@ class CpxE4AoUI(_CpxEModule):
         * True: Linear scaled
         """
         function_number = 4828 + 64 * self.position + 6
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register
         if value:
@@ -894,7 +894,7 @@ class CpxE4AoUI(_CpxEModule):
 
         """
         function_number = 4828 + 64 * self.position + 6
-        reg = self.base.read_function_number(function_number)[0]
+        reg = self.base.read_function_number(function_number)
 
         # Fill in the unchanged values from the register
         if value: